requirements.txt
fastapi
uvicorn
pydantic
requests
python-decouple
pandas
